Factored typeGenerator out into typeBuilder
